fix(driver): restore document expiry eligibility resolvers for compile
Restore DriverDocumentExpiryResolver and related governance files that were
left untracked outside main, unblocking 11.1.12+36 release AAB builds.

+++ admin/mndob-main/lib/backend/schema/order_record.dart
@@ -376,7 +376,7 @@ class OrderRecord extends FirestoreRecord {
 }
 
 Map<String, dynamic> createOrderRecordData({
-  int? total,
+  double? total,
   DocumentReference? user,
   DocumentReference? mndobUser,
   String? halhText,
